requirements.txt
pandas
numpy
matplotlib
folium
requests
PyYAML
openpyxl
konlpy
wordcloud
scipy
scikit-learn
